backend/requirements.txt
Flask==2.3.2
requests==2.31.0
m3u8==0.9.0
ffmpeg-python==0.2.0
